Exercícios - semana 02

+++ Módulo 3/m03s01-exercícios/src/main/java/tech/devinhouse/veiculos/service/VeiculoService.java
@@ -1,5 +1,6 @@
 package tech.devinhouse.veiculos.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tech.devinhouse.veiculos.exception.FalhaExclusaoVeiculoComMultasException;
@@ -9,8 +10,10 @@ import tech.devinhouse.veiculos.model.Veiculo;
 import tech.devinhouse.veiculos.repository.VeiculoRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
+@Slf4j
 public class VeiculoService {
 
     @Autowired
@@ -18,8 +21,10 @@ public class VeiculoService {
 
     public Veiculo inserir(Veiculo veiculo) {
         boolean existe = repo.existsById(veiculo.getPlaca());
-        if (existe)
+        if (existe) {
+            log.error("Veículo com placa já cadastrada: {}", veiculo.getPlaca());
             throw new RegistroExistenteException();
+        }
         veiculo = repo.save(veiculo);
         return veiculo;
     }
@@ -29,21 +34,35 @@ public class VeiculoService {
     }
 
     public Veiculo obter(String placa) {
-        return repo.findById(placa)
-                .orElseThrow(RegistroNaoEncontradoException::new);
+        Optional<Veiculo> veiculoOpt = repo.findById(placa);
+        if (veiculoOpt.isEmpty()) {
+            log.error("Veiculo de placa {} nao encontrado na base de dados!", placa);
+            throw new RegistroNaoEncontradoException();
+        }
+        return veiculoOpt.get();
     }
 
     public void excluir(String placa) {
-        Veiculo veiculo = repo.findById(placa)
-                .orElseThrow(RegistroNaoEncontradoException::new);
-        if (veiculo.getQtdMultas() != 0)
+        Optional<Veiculo> veiculoOpt = repo.findById(placa);
+        if (veiculoOpt.isEmpty()) {
+            log.error("Veiculo de placa {} nao encontrado na base de dados!", placa);
+            throw new RegistroNaoEncontradoException();
+        }
+        Veiculo veiculo = veiculoOpt.get();
+        if (veiculo.getQtdMultas() != 0) {
+            log.error("Veiculo com multas e portanto nao pode ser excluído!");
             throw new FalhaExclusaoVeiculoComMultasException();
+        }
         repo.deleteById(placa);
     }
 
     public Veiculo adicionarMulta(String placa) {
-        Veiculo veiculo = repo.findById(placa)
-                .orElseThrow(RegistroNaoEncontradoException::new);
+        Optional<Veiculo> veiculoOpt = repo.findById(placa);
+        if (veiculoOpt.isEmpty()) {
+            log.error("Veiculo de placa {} nao encontrado na base de dados!", placa);
+            throw new RegistroNaoEncontradoException();
+        }
+        Veiculo veiculo = veiculoOpt.get();
         int qtd = veiculo.getQtdMultas() + 1;
         veiculo.setQtdMultas(qtd);
         veiculo = repo.save(veiculo);
